- r2931-r2953
Useful regexen ran (regexes regex for easy searching later):
 - Diamondo25-style line endings
	Find {.*\)} {~(\{)}$
	Replace \1\2
 - Spaces after semicolon
	Find {.*\;}( |\t)+$
	Replace \1
 - Spaces after opening braces
	Find {.*\{}( |\t)+$
	Replace \1
 - Spaces after closing braces
	Find {.*\}}( |\t)+$
	Replace \1
 - Lines that are only spaces or tabs
	Find ^( |\t)+$
	Replace
 - Lines that begin with spaces instead of tabs
	Find ^{( )+}{.*}$
	Replace \2
 - Spaces after single-line comments
	Find {// .*}( |\t)+$
	Replace \1

git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/branches/buis_vana_overhaul@3065 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/ChannelServer.cpp
@@ -69,7 +69,7 @@ void ChannelServer::loadLogConfig() {
 }
 
 string ChannelServer::makeLogIdentifier() {
-	return "World: " + boost::lexical_cast<string>(getWorld()) + "; ID: " + boost::lexical_cast<string>(getChannel());
+	return "World: " + boost::lexical_cast<string>(static_cast<int16_t>(getWorld())) + "; ID: " + boost::lexical_cast<string>(getChannel());
 }
